Moved separated files into uti subdirectory
