+++ uti/utiByteIterator.hpp
