+++ uti/utiCharIterator.hpp
