+++ uti/utiCharIterator.inl
