+++ uti/utiReverseIterator.inl
