+++ uti/utiTest/utiTest.vcxproj.filters
@@ -13,6 +13,9 @@
       <UniqueIdentifier>{67DA6AB6-F800-4c08-8B7A-83BB121AAD01}</UniqueIdentifier>
       <Extensions>rc;ico;cur;bmp;dlg;rc2;rct;bin;rgs;gif;jpg;jpeg;jpe;resx;tiff;tif;png;wav;mfcribbon-ms</Extensions>
     </Filter>
+    <Filter Include="Header Files\uti">
+      <UniqueIdentifier>{6631ea47-d952-4635-9084-8e6848c17fba}</UniqueIdentifier>
+    </Filter>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="stdafx.h">
@@ -24,29 +27,29 @@
     <ClInclude Include="..\..\uti.hpp">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\utiAllocator.hpp">
-      <Filter>Header Files</Filter>
+    <ClInclude Include="..\utiAllocator.hpp">
+      <Filter>Header Files\uti</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\utiCommonHeader.hpp">
-      <Filter>Header Files</Filter>
+    <ClInclude Include="..\utiByteIterator.hpp">
+      <Filter>Header Files\uti</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\utiReferenceCounted.hpp">
-      <Filter>Header Files</Filter>
+    <ClInclude Include="..\utiCharIterator.hpp">
+      <Filter>Header Files\uti</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\utiByteIterator.hpp">
-      <Filter>Header Files</Filter>
+    <ClInclude Include="..\utiCommonHeader.hpp">
+      <Filter>Header Files\uti</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\utiCharIterator.hpp">
-      <Filter>Header Files</Filter>
+    <ClInclude Include="..\utiReferenceCounted.hpp">
+      <Filter>Header Files\uti</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\utiReverseIterator.hpp">
-      <Filter>Header Files</Filter>
+    <ClInclude Include="..\utiReverseIterator.hpp">
+      <Filter>Header Files\uti</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\utiUTF8String.hpp">
-      <Filter>Header Files</Filter>
+    <ClInclude Include="..\utiUTF8String.hpp">
+      <Filter>Header Files\uti</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\utiUTF16String.hpp">
-      <Filter>Header Files</Filter>
+    <ClInclude Include="..\utiUTF16String.hpp">
+      <Filter>Header Files\uti</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
@@ -67,26 +70,26 @@
     <Text Include="test.txt" />
   </ItemGroup>
   <ItemGroup>
-    <None Include="..\..\utiAllocator.inl">
-      <Filter>Header Files</Filter>
+    <None Include="..\utiAllocator.inl">
+      <Filter>Header Files\uti</Filter>
     </None>
-    <None Include="..\..\utiReferenceCounted.inl">
-      <Filter>Header Files</Filter>
+    <None Include="..\utiByteIterator.inl">
+      <Filter>Header Files\uti</Filter>
     </None>
-    <None Include="..\..\utiByteIterator.inl">
-      <Filter>Header Files</Filter>
+    <None Include="..\utiCharIterator.inl">
+      <Filter>Header Files\uti</Filter>
     </None>
-    <None Include="..\..\utiCharIterator.inl">
-      <Filter>Header Files</Filter>
+    <None Include="..\utiReferenceCounted.inl">
+      <Filter>Header Files\uti</Filter>
     </None>
-    <None Include="..\..\utiReverseIterator.inl">
-      <Filter>Header Files</Filter>
+    <None Include="..\utiReverseIterator.inl">
+      <Filter>Header Files\uti</Filter>
     </None>
-    <None Include="..\..\utiUTF8String.inl">
-      <Filter>Header Files</Filter>
+    <None Include="..\utiUTF8String.inl">
+      <Filter>Header Files\uti</Filter>
     </None>
-    <None Include="..\..\utiUTF16String.inl">
-      <Filter>Header Files</Filter>
+    <None Include="..\utiUTF16String.inl">
+      <Filter>Header Files\uti</Filter>
     </None>
   </ItemGroup>
 </Project>
